+++ sample-app/src/main/java/ca/antonious/sample/viewcells/EmptyViewCell.java
@@ -4,7 +4,7 @@ import android.view.View;
 import android.widget.TextView;
 
 import ca.antonious.sample.R;
-import ca.antonious.viewcelladapter.BaseViewHolder;
+import ca.antonious.viewcelladapter.viewcells.BaseViewHolder;
 import ca.antonious.viewcelladapter.viewcells.GenericViewCell;
 
 /**
