+++ sample-app/src/main/java/ca/antonious/sample/viewcells/HeaderViewCell.java
@@ -4,8 +4,7 @@ import android.view.View;
 import android.widget.TextView;
 
 import ca.antonious.sample.R;
-import ca.antonious.viewcelladapter.BaseViewHolder;
-import ca.antonious.viewcelladapter.annotations.BindListener;
+import ca.antonious.viewcelladapter.viewcells.BaseViewHolder;
 import ca.antonious.viewcelladapter.viewcells.GenericViewCell;
 
 /**
