+++ sample-app/src/main/java/ca/antonious/sample/viewcells/TaskViewCell.java
@@ -6,7 +6,7 @@ import android.widget.TextView;
 
 import ca.antonious.sample.R;
 import ca.antonious.sample.models.Task;
-import ca.antonious.viewcelladapter.BaseViewHolder;
+import ca.antonious.viewcelladapter.viewcells.BaseViewHolder;
 import ca.antonious.viewcelladapter.annotations.BindListener;
 import ca.antonious.viewcelladapter.viewcells.GenericViewCell;
