+++ viewcelladapter/src/main/java/ca/antonious/viewcelladapter/viewcells/GenericViewCell.java
@@ -2,9 +2,6 @@ package ca.antonious.viewcelladapter.viewcells;
 
 import java.lang.reflect.ParameterizedType;
 
-import ca.antonious.viewcelladapter.BaseViewHolder;
-import ca.antonious.viewcelladapter.viewcells.ViewCell;
-
 /**
  * Created by George on 2016-11-17.
  */
