+++ viewcelladapter/src/test/java/ca/antonious/viewcelladapter/TestViewCell.java
@@ -2,7 +2,7 @@ package ca.antonious.viewcelladapter;
 
 import android.view.View;
 
-import ca.antonious.viewcelladapter.viewcells.AbstractViewCell;
+import ca.antonious.viewcelladapter.viewcells.BaseViewHolder;
 import ca.antonious.viewcelladapter.viewcells.GenericViewCell;
 
 /**
